refactor: moved createstorage function to common module
